authentication part

+++ account/views.py
@@ -8,6 +8,8 @@ from rest_framework.generics import UpdateAPIView
 from rest_framework.response import Response
 from django.contrib.auth import authenticate, login, logout
 from rest_framework.authtoken.models import Token
+from rest_framework.permissions import IsAuthenticated
+
 
 from django.contrib.auth.tokens import default_token_generator
 from django.utils.http import urlsafe_base64_encode, urlsafe_base64_decode
@@ -83,8 +85,10 @@ class UserLoginApiView(APIView):
         return Response(serializer.errors)
     
 class UserLogoutApiView(APIView):
+    permission_classes = [IsAuthenticated]
     def get(self, request):
-        request.user.auth_token.delete()
+        if request.user.auth_token:
+            request.user.auth_token.delete()
         logout(request)
         return redirect('login')
